Allow automating instance recovery on system/instance check failure
This feature depends on instance store volumes not being attached to the EC2
instance, for reasons internal to AWS EC2. Those can be removed using the AWS
CLI or with this module. However, doing the removal with this module seems to
trigger a rebuild of the EC2 instance for no good reason.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.net>